add LogoLink component and rename renderLogo to showCredit
