+++ packages/pyroscope-flamegraph/src/FlameGraph/FlamegraphRenderer.module.scss
@@ -36,3 +36,11 @@
 .tablePane {
   flex: 1;
 }
+
+.logoLink {
+  display: inline-block;
+  text-decoration: none;
+  font-weight: 500;
+  opacity: 0.5;
+  color: var(--ps-neutral-2);
+}
